JavaScript (v3): Onboard some SQS guide entries into SOS.
