+++ javascriptv3/example_code/sqs/actions/receive-delete-message.js
@@ -23,7 +23,6 @@ const receiveMessage = (queueUrl) =>
       MessageAttributeNames: ["All"],
       QueueUrl: queueUrl,
       VisibilityTimeout: 20,
-      WaitTimeSeconds: 5,
     })
   );
