+++ javascriptv3/example_code/sqs/actions/set-attributes-dead-letter.js
@@ -1,0 +1,42 @@
+/*
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { fileURLToPath } from "url";
+
+// snippet-start:[sqs.JavaScript.deadLetter.setQueueAttributesV3]
+import { SetQueueAttributesCommand, SQSClient } from "@aws-sdk/client-sqs";
+
+const client = new SQSClient({});
+const SQS_QUEUE_URL = "queue_url";
+const DEAD_LETTER_QUEUE_ARN = "dead_letter_queue_arn";
+
+export const main = async (
+  queueUrl = SQS_QUEUE_URL,
+  deadLetterQueueArn = DEAD_LETTER_QUEUE_ARN
+) => {
+  const command = new SetQueueAttributesCommand({
+    Attributes: {
+      RedrivePolicy: JSON.stringify({
+        // Amazon SQS supports dead-letter queues (DLQ), which other
+        // queues (source queues) can target for messages that can't
+        // be processed (consumed) successfully.
+        // https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-dead-letter-queues.html
+        deadLetterTargetArn: deadLetterQueueArn,
+        maxReceiveCount: "10",
+      }),
+    },
+    QueueUrl: queueUrl,
+  });
+
+  const response = await client.send(command);
+  console.log(response);
+  return response;
+};
+// snippet-end:[sqs.JavaScript.deadLetter.setQueueAttributesV3]
+
+// Invoke main function if this file was run directly.
+if (process.argv[1] === fileURLToPath(import.meta.url)) {
+  main();
+}
