--- javascriptv3/example_code/sqs/guide_supplements/guide-js-update-queue-long-polling.js
@@ -1,36 +1,0 @@
-/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-/**
- * This example supplements the following guide. For more context, see
- * https://docs.aws.amazon.com/sdk-for-javascript/v3/developer-guide/sqs-examples-enable-long-polling.html
- */
-
-import { fileURLToPath } from "url";
-
-// snippet-start:[sqs.JavaScript.longPoll.setQueueAttributesV3]
-import { SetQueueAttributesCommand, SQSClient } from "@aws-sdk/client-sqs";
-
-const client = new SQSClient({});
-const SQS_QUEUE_URL = "queue_url";
-
-export const main = async (queueUrl = SQS_QUEUE_URL) => {
-  const command = new SetQueueAttributesCommand({
-    Attributes: {
-      ReceiveMessageWaitTimeSeconds: "20",
-    },
-    QueueUrl: queueUrl,
-  });
-
-  const response = await client.send(command);
-  console.log(response);
-  return response;
-};
-// snippet-end:[sqs.JavaScript.longPoll.setQueueAttributesV3]
-
-// Invoke main function if this file was run directly.
-if (process.argv[1] === fileURLToPath(import.meta.url)) {
-  main();
-}
